// Client/src/app/chart-of-accounts.ts
export class CoA {
    caId: number;
    accountType: string;
    accountSubType: string;
    reportType: string;
    accountNumber: number;
    normalSide: string;
    originalBalance: number;
    currentBalance: number;
    active: string;
    comment: string;
    accountName: string;
    createdBy: string;
  }
  